move to generator-common
