+++ packages/generator/src/file-generator.ts
@@ -1,7 +1,8 @@
 import { readFileSync } from 'fs';
 import { resolve } from 'path';
-import { codeBlock, createLogger } from '@sap-cloud-sdk/util';
+import { createLogger } from '@sap-cloud-sdk/util';
 import { Directory, SourceFile, SourceFileStructure } from 'ts-morph';
+import { getCopyrightHeader } from '@sap-cloud-sdk/generator-common/internal';
 
 const logger = createLogger({
   package: 'generator',
@@ -64,15 +65,4 @@ export function copyFile(
 function addFileComment(content: SourceFileStructure): SourceFileStructure {
   content.leadingTrivia = getCopyrightHeader();
   return content;
-}
-
-// TODO 1728 move to a new package for reduce code duplication.
-function getCopyrightHeader(): string {
-  return codeBlock`
-/*
- * Copyright (c) ${new Date().getFullYear()} SAP SE or an SAP affiliate company. All rights reserved.
- *
- * This is a generated file powered by the SAP Cloud SDK for JavaScript.
- */
- `;
 }
