--- packages/openapi-generator/src/parser/reserved-words.ts
@@ -1,47 +1,0 @@
-// TODO 1728: The following is duplicate in the openapi generator
-/**
- * @internal
- */
-export const reservedJsKeywords: readonly string[] = [
-  'break',
-  'case',
-  'catch',
-  'class',
-  'const',
-  'continue',
-  'debugger',
-  'default',
-  'delete',
-  'do',
-  'else',
-  'enum',
-  'export',
-  'extends',
-  'false',
-  'finally',
-  'for',
-  'function',
-  'if',
-  'implements',
-  'in',
-  'instanceof',
-  'let',
-  'new',
-  'null',
-  'protected',
-  'public',
-  'return',
-  'static',
-  'super',
-  'switch',
-  'symbol',
-  'this',
-  'true',
-  'try',
-  'typeof',
-  'var',
-  'void',
-  'while',
-  'with',
-  'yield'
-] as const;
